requirements.txt
# Core dependencies
aiohttp>=3.10.10
python-dotenv==1.0.1
async-timeout<5.0,>=4.0
requests>=2.32.3
fastapi>=0.115.6
uvicorn[standard]>=0.25.0
websockets>=10.0
tenacity>=8.5.0
rich>=13.0.0

# Database drivers
neo4j>=5.14.0
motor>=3.3.2
pytz>=2024.2
pydantic>=2.10.6

# AI/ML packages
openai>=1.12.0
google-generativeai>=0.8.3
together>=0.2.0
anthropic>=0.37.1
voyageai>=0.3.2
groq>=0.13.1

# LangChain and related packages
langchain-google-genai>=2.0.1
langchain>=0.3.4
langchain-core>=0.3.12
langchain-experimental>=0.0.49
langchain-community>=0.0.13
langchain-openai>=0.2.3
langchain-anthropic>=0.2.3

# Message processing
tiktoken>=0.5.2

# NEAR Protocol integration
borsh-construct==0.1.0
base58==2.1.1
pynacl>=1.5.0
py-near>=0.2.1  # Replacing near-api>=0.3.0 with the recommended package